requirements.txt
pillow~=12.1.0
pycocotools~=2.0.11
matplotlib
